requirements.txt
streamlit
Python-dotenv
google.generativeai
SpeechRecognition
pyaudio
gTTS 
pygame